requirements.txt
aiohttp==3.9.3
aiofiles==23.2.1
python-dotenv==1.0.1
fastapi==0.109.2
uvicorn==0.27.1
jinja2==3.1.3
websockets==12.0
aiosqlite==0.19.0 